Task 6

+++ 0x00-hello_world/6-size.c
@@ -13,7 +13,7 @@ int main(void) /* main block returns the value 0 */
 	float floatType;
 	char charType;
 	long longType;
-	double doubleType;
+	long long doubleType;
 
 /* sizeof evaluates the value of a variable */
 	printf("The size of an int : %zu byte(s)\n", sizeof(intType));
